move column selection based on required_variables parameter to the base refuter class (#128)
* move column selection to the base class

* refactor code to merge functionality of sanity check and choosing_variables

* make invert a local variable

* remove call to sanity check

* fix typo in causal_refuter
