+++ tests/causal_refuters/test_bootstrap_refuter.py
@@ -11,20 +11,20 @@ class TestDataSubsetRefuter(object):
     '''
 
     @pytest.mark.parametrize(["error_tolerance","estimator_method"],
-                              [(0.01, "iv.instrumental_variable")])
+                              [(0.05, "iv.instrumental_variable")])
     def test_refutation_bootstrap_refuter_continuous(self, error_tolerance, estimator_method):
         refuter_tester = TestRefuter(error_tolerance, estimator_method, "bootstrap_refuter")
         refuter_tester.continuous_treatment_testsuite() # Run both
 
     @pytest.mark.parametrize(["error_tolerance", "estimator_method"],
-                             [(0.01, "backdoor.propensity_score_matching")])
+                             [(0.05, "backdoor.propensity_score_matching")])
     def test_refutation_bootstrap_refuter_binary(self, error_tolerance, estimator_method):
         refuter_tester = TestRefuter(error_tolerance, estimator_method, "bootstrap_refuter")
         refuter_tester.binary_treatment_testsuite(tests_to_run="atleast-one-common-cause")
 
     
     @pytest.mark.parametrize(["error_tolerance","estimator_method","num_common_causes","required_variables"],
-                              [(0.01, "iv.instrumental_variable",5, 3)])
+                              [(0.05, "iv.instrumental_variable",5, 3)])
     def test_refutation_bootstrap_refuter_continuous_integer_argument(self, error_tolerance, estimator_method, num_common_causes, required_variables):
         refuter_tester = TestRefuter(error_tolerance, 
                                      estimator_method, 
@@ -34,7 +34,7 @@ class TestDataSubsetRefuter(object):
         refuter_tester.continuous_treatment_testsuite(num_common_causes=num_common_causes, tests_to_run="atleast-one-common-cause") # Run atleast one common cause
 
     @pytest.mark.parametrize(["error_tolerance","estimator_method", "num_common_causes", "required_variables"],
-                              [(0.01, "iv.instrumental_variable", 5, ["W0","W1"])])
+                              [(0.05, "iv.instrumental_variable", 5, ["W0","W1"])])
     def test_refutation_bootstrap_refuter_continuous_list_argument(self, error_tolerance, estimator_method, num_common_causes, required_variables):
         refuter_tester = TestRefuter(error_tolerance,
                                      estimator_method,
@@ -61,7 +61,7 @@ class TestDataSubsetRefuter(object):
         refuter_tester.binary_treatment_testsuite(num_common_causes=num_common_causes, tests_to_run="atleast-one-common-cause")
 
     @pytest.mark.parametrize(["error_tolerance","estimator_method", "num_common_causes", "required_variables"],
-                              [(0.01, "iv.instrumental_variable", 5, ["-W0","-W1"])])
+                              [(0.05, "iv.instrumental_variable", 5, ["-W0","-W1"])])
     def test_refutation_bootstrap_refuter_continuous_list_negative_argument(self, error_tolerance, estimator_method, num_common_causes, required_variables):
         refuter_tester = TestRefuter(error_tolerance,
                                      estimator_method,
